completed Fortran kind definition changes for Fortran test drivers

+++ branches/p3dfft-k/sample/FORTRAN/driver_sine.f90
@@ -36,11 +36,11 @@
       complex(mytype), dimension(:,:,:),  allocatable :: AEND
       real(mytype) pi,twopi,sinyz,diff,cdiff,ccdiff,ans
 
-      integer(8) Ntot
+      integer(i8) Ntot
       real(mytype) factor
       real(mytype),dimension(:),allocatable:: sinx,siny,sinz
-      real(8) rtime1,rtime2,Nglob
-      real(8) gt(10,3),gtcomm(3),tc
+      real(r8) rtime1,rtime2,Nglob
+      real(r8) gt(10,3),gtcomm(3),tc
       integer ierr,nu,ndim,dims(2),nproc,proc_id
       integer istart(3),iend(3),isize(3)
       integer fstart(3),fend(3),fsize(3)
@@ -283,7 +283,7 @@
 
       use p3dfft
 
-      integer(8) nar,i
+      integer(i8) nar,i
       complex(mytype) X(nar)
       real(mytype) f
 
@@ -304,10 +304,10 @@
       use p3dfft
 
       integer x,y,z,proc_id
-      integer(8) i,Nar
+      integer(i8) i,Nar
       complex(mytype) Ar(1,1,*)
       integer Fstart(3),Fend(3),Fsize(3)
-      real(8) Nglob
+      real(r8) Nglob
 
       call p3dfft_get_dims(Fstart,Fend,Fsize,2)
